Initialize names slice to the number of entries.
Also, do not call os.RemoveAll() unless the path to the directory
is a valid one.

+++ caddyhttp/httpserver/context_test.go
@@ -772,8 +772,10 @@ func TestFiles(t *testing.T) {
 			}
 		}
 
-		if err := os.RemoveAll(dirPath); err != nil && !os.IsNotExist(err) {
-			t.Fatalf(testPrefix+"Expected no error removing directory, got: '%s'", err.Error())
+		if dirPath != "" {
+			if err := os.RemoveAll(dirPath); err != nil && !os.IsNotExist(err) {
+				t.Fatalf(testPrefix+"Expected no error removing directory, got: '%s'", err.Error())
+			}
 		}
 	}
 }
